Moved ratings fetch to the login component
