+++ src/App/App.js
@@ -61,7 +61,7 @@ class App extends Component {
     })
   }
 
-  appState = (userData) => {   
+  appState = (userData) => { 
     this.setState({ 
       userInfo: {
         userId: userData.userId ,
@@ -69,17 +69,17 @@ class App extends Component {
         userRatings: userData.userRatings
       }
     })
+    console.log(this.state.userInfo);
+    
   }
 
   render() {  
-    console.log('app rendered');
-    
     return this.state.movies ?
     <main className="App">
       <Route path = '/' render = { () =>  <Header onClick={ this.clickHandler } appState={ this.state } logoutButton={ this.userLoggedIn} /> }/>
-      <Route exact path = '/' render = { () =>  <Movies movies={ this.state.movies} onClick={ this.clickHandler } appState={ this.state } updateAppState={ this.appState } /> }/>
+      <Route exact path = '/' render = { () =>  <Movies movies={ this.state.movies} onClick={ this.clickHandler } appState={ this.state } /> }/>
       <Route exact path = '/login' render =  { () => <LogInPage userInfo={ this.state.userInfo } updateAppState={ this.userLoggedIn } appState={ this.appState }/> }/>
-      <Route exact path = '/users/63' render =  { () => <Movies movies={ this.state.movies} onClick={ this.clickHandler } appState={ this.state } updateAppState={ this.appState }/> }/>
+      <Route exact path = '/users/63' render =  { () => <Movies movies={ this.state.movies} onClick={ this.clickHandler } appState={ this.state } /> }/>
       <Route exact path = '/movie/:id' render =  { ({match}) => <ExpandedMovie {...match} movieDoneLoading= { this.movieDidLoad } appState = { this.state }/> } />
     </main>
     :
