+++ src/FetchedData/FetchedData.js
@@ -36,7 +36,7 @@ export const postLogin = async (email, password) => {
 
 export const getUserRatings = async () => {
     //Change this back to the correct user
-    const response = await fetch('https://rancid-tomatillos.herokuapp.com/api/v2/users/60/ratings')
+    const response = await fetch('https://rancid-tomatillos.herokuapp.com/api/v2/users/62/ratings')
     const userRatings = await response.json()
     return userRatings;
